+++ packages/core-js/modules/esnext.global-this.js
@@ -1,8 +1,2 @@
-var $ = require('../internals/export');
-var global = require('../internals/global');
-
-// `globalThis` object
-// https://github.com/tc39/proposal-global
-$({ global: true }, {
-  globalThis: global
-});
+// TODO: Remove from `core-js@4`
+require('./es.global-this');