+++ packages/core-js/stable/global-this.js
@@ -1,3 +1,1 @@
-require('../modules/esnext.global-this');
-
-module.exports = require('../internals/path').globalThis;
+module.exports = require('../es/global-this');