+++ packages/core-js/stage/3.js
@@ -1,3 +1,1 @@
-require('../proposals/global-this');
-
 module.exports = require('./4');